+++ spec/head_music/notation/lily_pond_round_trip_spec.rb
@@ -41,6 +41,7 @@ module LilyPondRoundTripSources
 
   FIXTURE_NAMES = %i[
     speed_the_plough chromatic_air rests duo key_and_meter_change tacet song escaped_header anonymous air fourth_species
+    tie_into_key_change tie_into_meter_change
   ].freeze
 
   ALL_HAND_WRITTEN = RELATIVE_TWINS.flat_map { |name, twin| [["#{name} (relative)", twin[:relative]], ["#{name} (absolute)", twin[:absolute]]] }